fix identation and use mixins to reduce too many lines in ReportsController
