+++ app/controllers/reports_controller.rb
@@ -1,6 +1,6 @@
 class ReportsController < ApplicationController
   before_action :set_report, only: [:show, :edit, :update, :destroy]
-
+  include ReportHelper
   # GET /reports
   # GET /reports.json
   def index
@@ -10,10 +10,10 @@ class ReportsController < ApplicationController
   # GET /reports/1
   # GET /reports/1.json
   def show
-    @report_details = ReportsController.report_details(@report)
-    @view_level = params[:view_level] ? params[:view_level] : 'genus'
-    @taxonomy_details = ReportsController.taxonomy_details(@view_level, @report, params)
-    @highest_tax_counts = ReportsController.highest_tax_counts(@view_level, @report)
+    @report_details = report_details(@report)
+    @view_level = params[:view_level] ? params[:view_level].downcase : 'genus'
+    @taxonomy_details = taxonomy_details(@view_level, @report, params)
+    @highest_tax_counts = highest_tax_counts(@view_level, @report)
   end
 
   # GET /reports/new
@@ -70,80 +70,6 @@ class ReportsController < ApplicationController
     respond_to do |format|
       format.html { redirect_to reports_url, notice: 'Report was successfully destroyed.' }
       format.json { head :no_content }
-    end
-  end
-
-  class << self
-    def external_report_info(report, view_level, params)
-      data = {}
-      data[:report_details] = ReportsController.report_details(report)
-      data[:taxonomy_details] = ReportsController.taxonomy_details(view_level, report, params)
-      data[:highest_tax_counts] = ReportsController.highest_tax_counts(view_level, report)
-      data[:view_level] = view_level
-      data
-    end
-
-    def report_details(report)
-      {
-        report_info: report,
-        pipeline_info: report.pipeline_output,
-        sample_info: report.pipeline_output.sample,
-        project_info: report.pipeline_output.sample.project,
-        background_model: report.background
-      }
-    end
-
-    def taxonomy_details(view_level, report, params)
-      view_level = view_level.downcase
-      zscores = report.taxon_zscores
-      nt_zscore_threshold = params[:nt_zscore_threshold]
-      nt_rpm_threshold = params[:nt_rpm_threshold]
-      nr_zscore_threshold = params[:nr_zscore_threshold]
-      nr_rpm_threshold = params[:nr_rpm_threshold]
-      tax_details = []
-      if view_level == 'species'
-        nt_zscores = zscores.type('NT').level(TaxonCount::TAX_LEVEL_SPECIES)
-        nr_zscores = zscores.type('NR').level(TaxonCount::TAX_LEVEL_SPECIES)
-      else
-        nt_zscores = zscores.type('NT').level(TaxonCount::TAX_LEVEL_GENUS)
-        nr_zscores = zscores.type('NR').level(TaxonCount::TAX_LEVEL_GENUS)
-      end
-      # TODO, discuss why the NT score is sorted and not the NR score
-      tax_ids = filter(nt_zscores, nt_zscore_threshold, nr_zscore_threshold,
-                       nt_rpm_threshold, nr_rpm_threshold)
-                .order(zscore: :desc).where.not("tax_id < 0").limit(20).map(&:tax_id)
-      tax_ids.each do |id|
-        nt_ele = nt_zscores.find_by(tax_id: id)
-        nr_ele = nr_zscores.find_by(tax_id: id)
-        tax_details.push(nt_ele: nt_ele, nr_ele: nr_ele)
-      end
-      tax_details
-    end
-
-    def highest_tax_counts(view_level, report)
-      view_level = view_level.downcase
-      zscores = report.taxon_zscores
-      metrics = {}
-      if view_level == 'species'
-        nt_zscores = zscores.type('NT').level(TaxonCount::TAX_LEVEL_SPECIES)
-      else
-        nt_zscores = zscores.type('NT').level(TaxonCount::TAX_LEVEL_GENUS)
-      end
-      highest_zscore = nt_zscores.order(zscore: :desc).first
-      highest_rpm = nt_zscores.order(rpm: :desc).first
-      if highest_zscore && highest_rpm
-        metrics[:highest_zscore] = highest_zscore[:zscore]
-        metrics[:highest_rpm] = highest_rpm[:rpm]
-      end
-      metrics
-    end
-
-    def filter(nt_zscores, nt_threshold, nr_threshold, nt_rpm_threshold, nr_rpm_threshold)
-      nt_zscores = nt_zscores.where('zscore >= ?', nt_threshold) if nt_threshold
-      nt_zscores = nt_zscores.where('zscore >= ?', nr_threshold) if nr_threshold
-      nt_zscores = nt_zscores.where('rpm >= ?', nt_rpm_threshold) if nt_rpm_threshold
-      nt_zscores = nt_zscores.where('rpm >= ?', nr_rpm_threshold) if nr_rpm_threshold
-      nt_zscores
     end
   end
